tenth commit

+++ src/main/java/com/onelshina/csis/Fibonacci.java
@@ -6,7 +6,7 @@ public class Fibonacci {
 
     public static void main(String[] args) {
         long startingTime = System.currentTimeMillis();
-        long n = 2000; //Change this number to get the Fibonacci number you want
+        int n = 2000; //Change this number to get the Fibonacci number you want
         String[] result = calcFib(n);
         //Index = 0 : Fibonacci Number
         //Index = 1 : Count (How many times recursive method was called)
@@ -19,22 +19,22 @@ public class Fibonacci {
     }
 
     /**
-     * Calls {@link #calcFibAux(long, BigInteger, BigInteger, long)}  method that
+     * Calls {@link #calcFibAux(int, BigInteger, BigInteger, int)}  method that
      * returns a String Array including the Fibonacci number at index 0, Count at index 1.
      *
      * This method also passes values required to start the recursive calls
-     * into {@link #calcFibAux(long, BigInteger, BigInteger, long)}
+     * into {@link #calcFibAux(int, BigInteger, BigInteger, int)}
      *
-     * @param n {@link long} The number we want to find the Fibonacci Number for
+     * @param n {@link int} The number we want to find the Fibonacci Number for
      * @return {@link String[]} (index = 0: Fibonacci number
      * index = 1: count "how many times recursion method was called")
      */
-    protected static String[] calcFib(long n) {
+    protected static String[] calcFib(int n) {
        return calcFibAux(n, BigInteger.ZERO, BigInteger.ONE, 0);
     }
 
 
-    private static String[] calcFibAux(long n, BigInteger first, BigInteger second, long count) {
+    private static String[] calcFibAux(int n, BigInteger first, BigInteger second, int count) {
         if (n == 0) { //Base Case
             return new String[]{String.valueOf(first), String.valueOf(count)};
         } else {
